+++ src/assets/ba_data/python/bauiv1lib/characterpicker.py
@@ -198,7 +198,7 @@ class CharacterPicker(PopupWindow):
         plus = bui.app.plus
         assert plus is not None
 
-        if plus.get_v1_account_state() != 'signed_in':
+        if plus.accounts.primary is None:
             show_sign_in_prompt()
             return