+++ src/assets/ba_data/python/bauiv1lib/docui/_controller.py
@@ -295,7 +295,7 @@ class DocUIController:
         origin_widget: bui.Widget | None = None,
         auxiliary_style: bool = True,
         uiopenstateid: str | None = None,
-        suppress_ui_open_state_warning: bool = False,
+        suppress_win_extra_type_warning: bool = False,
     ) -> DocUIWindow:
         """Create a new window to handle a request."""
         assert bui.in_logic_thread()
@@ -308,7 +308,7 @@ class DocUIController:
             origin_widget=origin_widget,
             auxiliary_style=auxiliary_style,
             uiopenstateid=uiopenstateid,
-            suppress_ui_open_state_warning=suppress_ui_open_state_warning,
+            suppress_win_extra_type_warning=suppress_win_extra_type_warning,
         )
         self._set_win_state(win, self._WinState.FETCHING_FRESH_REQUEST)
 
@@ -527,7 +527,7 @@ class DocUIController:
                         action.request,
                         origin_widget=widget,
                         auxiliary_style=False,
-                        suppress_ui_open_state_warning=True,
+                        suppress_win_extra_type_warning=True,
                     )
                 )