+++ src/ballistica/shared/ballistica.h
@@ -151,7 +151,8 @@ enum class Permission : uint8_t {
 };
 
 // BA_EXPORT_PYTHON_ENUM
-/// Special characters the game can print.
+/// Special characters the engine can diplay. Note that this currently
+/// needs to be manually kept in sync with bacommon.text.SpecialChar.
 enum class SpecialChar : uint8_t {
   kDownArrow,
   kUpArrow,